ENH: Further organization.
