+++ Modules/Filtering/AdaptiveDenoising/include/itkNonLocalPatchBasedImageFilter.h
