+++ Modules/Filtering/AdaptiveDenoising/include/itkNonLocalPatchBasedImageFilter.hxx
