+++ Modules/Filtering/AdaptiveDenoising/include/itkVarianceImageFilter.h
