+++ Modules/Filtering/AdaptiveDenoising/include/itkVarianceImageFilter.hxx
