+++ Modules/Filtering/AdaptiveDenoising/itk-module.cmake
@@ -12,7 +12,9 @@ itk_module(
   AdaptiveDenoising
   DEPENDS
     ITKCommon
+    ITKSmoothing
     ITKStatistics
+    ITKImageStatistics
   COMPILE_DEPENDS
     ITKImageSources
   TEST_DEPENDS
