--- Modules/Filtering/AdaptiveDenoising/src/CMakeLists.txt
@@ -1,5 +1,0 @@
-set(AdaptiveDenoising_SRCS
-  itkMinimalStandardRandomVariateGenerator.cxx
-  )
-
-itk_module_add_library(AdaptiveDenoising ${AdaptiveDenoising_SRCS})
