--- Modules/Filtering/AdaptiveDenoising/test/Baseline/itkAdaptiveNonLocalMeansDenoisingImageFilterTestOutput.mha.sha512
@@ -1,1 +1,0 @@
-34cdd2aafce413dcf7efc692a707d53b7b57ebb006d0fbc20ecb42343f8a078b2252eff2137b17d3ab583d133f780438e247241ebf213c3289bd9ba1525081b7
