+++ Modules/Filtering/AdaptiveDenoising/test/CMakeLists.txt
@@ -1,27 +1,21 @@
 itk_module_test()
 
-set(
-  AdaptiveDenoisingTests
-  itkMinimalStandardRandomVariateGeneratorTest.cxx
-  itkAdaptiveNonLocalMeansDenoisingImageFilterTest.cxx
-)
+set(AdaptiveDenoisingTests itkAdaptiveNonLocalMeansDenoisingImageFilterTest.cxx)
 
 createtestdriver(AdaptiveDenoising "${AdaptiveDenoising-Test_LIBRARIES}" "${AdaptiveDenoisingTests}")
 
 itk_add_test(
-  NAME itkMinimalStandardRandomVariateGeneratorTest
+  NAME itkAdaptiveNonLocalMeansDenoisingImageFilterTest1
   COMMAND
     AdaptiveDenoisingTestDriver
-    itkMinimalStandardRandomVariateGeneratorTest
+    itkAdaptiveNonLocalMeansDenoisingImageFilterTest
 )
 
-itk_add_test(
-  NAME itkAdaptiveNonLocalMeansDenoisingImageFilterTest
-  COMMAND
-    AdaptiveDenoisingTestDriver
-    --compare
-    DATA{Baseline/itkAdaptiveNonLocalMeansDenoisingImageFilterTestOutput.mha}
-    ${ITK_TEST_OUTPUT_DIR}/itkAdaptiveNonLocalMeansDenoisingImageFilterTestOutput.mha
-    itkAdaptiveNonLocalMeansDenoisingImageFilterTest
-    ${ITK_TEST_OUTPUT_DIR}/itkAdaptiveNonLocalMeansDenoisingImageFilterTestOutput.mha
-)
+# itk_add_test(NAME itkAdaptiveNonLocalMeansDenoisingImageFilterTest
+#  COMMAND AdaptiveDenoisingTestDriver
+#    --compare
+#    DATA{Baseline/itkAdaptiveNonLocalMeansDenoisingImageFilterTestOutput.mha}
+#    ${ITK_TEST_OUTPUT_DIR}/itkAdaptiveNonLocalMeansDenoisingImageFilterTestOutput.mha
+#  itkAdaptiveNonLocalMeansDenoisingImageFilterTest
+#    ${ITK_TEST_OUTPUT_DIR}/itkAdaptiveNonLocalMeansDenoisingImageFilterTestOutput.mha
+# )
