+++ Modules/Filtering/AdaptiveDenoising/wrapping/CMakeLists.txt
@@ -1,3 +1,6 @@
 itk_wrap_module(AdaptiveDenoising)
+
+set(WRAPPER_SUBMODULE_ORDER itkAdaptiveNonLocalMeansDenoisingImageFilter)
+
 itk_auto_load_submodules()
 itk_end_wrap_module()
